[domain] move S3MetaData into module
