+++ domain/src/main/scala/net/kemitix/s3thorp/domain/S3MetaData.scala
@@ -1,6 +1,4 @@
-package net.kemitix.s3thorp
-
-import net.kemitix.s3thorp.domain.{LocalFile, RemoteMetaData}
+package net.kemitix.s3thorp.domain
 
 // For the LocalFile, the set of matching S3 objects with the same MD5Hash, and any S3 object with the same remote key
 final case class S3MetaData(
